docs(profile): document store interface methods
these will be changing. Best to outline them properly now
